refactor: split gateway/discord into service adapters and command handlers
