+++ gateway/discord/command/command_handler_test.go
@@ -1,4 +1,4 @@
-package discord
+package command
 
 import (
 	"testing"
